ui updates
- change primary color
- organize theme into theme.js
